+++ app/Http/Controllers/ResourceController.php
@@ -9,10 +9,13 @@ use App\ElasticDataSet;
 use Chumper\Zipper\Facades\Zipper;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Session;
 use App\Http\Controllers\Api\ResourceController as ApiResource;
 use App\Http\Controllers\Api\ConversionController as ApiConversion;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+use PhpParser\Node\Scalar\MagicConst\Dir;
 
 class ResourceController extends Controller
 {
@@ -137,7 +140,27 @@ class ResourceController extends Controller
         if (in_array($metadata['data']['type'], [Resource::TYPE_HYPERLINK, Resource::TYPE_AUTO])) {
           $success = true;
         } else if (!empty($extension)) {
-          $data = self::callConversions($apiKey, $extension, $content, $uri);
+          if(strtolower($extension) == "zip") {
+            $data['zip'] = "zip";
+
+            $zipDir = "../storage/app/files";
+            if(!file_exists($zipDir) && !is_dir($zipDir)) {
+              mkdir($zipDir, 0777);
+            }
+
+            $zipDir = "../storage/app/files/$uri";
+            if(!file_exists($zipDir) && !is_dir($zipDir)) {
+              \Log::info("Create folder $zipDir at {$_SERVER['SERVER_NAME']}");
+              mkdir($zipDir, 0777);
+            }
+
+            $file->move($zipDir, $file->getClientOriginalName());
+
+            $success = true;
+          }
+          else {
+            $data = self::callConversions($apiKey, $extension, $content, $uri);
+          }
         }
 
         if (Session::has('elasticData.'. $uri)) {
@@ -633,6 +656,44 @@ class ResourceController extends Controller
   }
 
   /**
+   * Download resource zip file
+   *
+   * @param Request $request - resource uri and file name
+   *
+   * @return downlodable file
+   */
+  public function resourceDirectDownloadZip(Request $request)
+  {
+    $uri = $request->offsetGet('uri');
+    $zipDir = "../storage/app/files/$uri";
+    $zipFile = glob($zipDir.DIRECTORY_SEPARATOR."*");
+    if(empty($zipFile)) {
+      $request->session()->flash('alert-danger', __('custom.zip_download_error'));
+      return redirect()->back()->withInput();
+    }
+    $zipNameExpl = explode("/", $zipFile[0]);
+    $zipName = end($zipNameExpl);
+
+    if(file_exists($zipDir.DIRECTORY_SEPARATOR.$zipName)) {
+      \Log::info("Reding file $zipName from folder $zipDir from server {$_SERVER['SERVER_NAME']}");
+
+      header('Content-Description: File Transfer');
+      header('Content-Type: application/zip');
+      header('Content-disposition: attachment; filename='.@basename($zipName));
+      header('Expires: 0');
+      header('Cache-Control: must-revalidate, post-check=0, pre-check=0');
+      header('Pragma: public');
+      header('Content-Length: ' .@filesize($zipDir.DIRECTORY_SEPARATOR.$zipName));
+
+      if(ob_get_level()){
+        ob_end_clean();
+      }
+
+      readfile($zipDir.DIRECTORY_SEPARATOR.$zipName);
+    }
+  }
+
+  /**
    * Transforms resource data to downloadable file
    *
    * @param Request $request - file format, uri of the resource
@@ -728,7 +789,19 @@ class ResourceController extends Controller
 
       try {
 
-        foreach ($resources as $resource) {
+        $thereIsResourceZip = false;
+        $zipResources = [];
+
+        foreach ($resources as $key => $resource) {
+
+          if($resource->file_format == Resource::getFormats()[Resource::FORMAT_ZIP]) {
+            $zipDirResource = "../storage/app/files/$resource->uri";
+            $zipFileName = Resource::getResourceZipFile($resource->uri);
+            $zipResources[$key]['dir'] = $zipDirResource;
+            $zipResources[$key]['name'] = $zipFileName;
+            $thereIsResourceZip = true;
+            continue;
+          }
 
           $resourceId = $resource->id;
           $checkName = (mb_strlen($resource->name) > 130) ? mb_substr($resource->name, 0, 130, "utf-8") : $resource->name;
@@ -777,6 +850,11 @@ class ResourceController extends Controller
               $exploded = explode("/", $file);
               $fileName = end($exploded);
               $zip->addFile($file, $fileName.".".strtolower($format));
+            }
+            if($thereIsResourceZip) {
+              foreach ($zipResources as $zipResource) {
+                $zip->addFile($zipResource['dir'].DIRECTORY_SEPARATOR.$zipResource['name'], $zipResource['name']);
+              }
             }
             $zip->close();
           }